requirements.txt
pandas
scikit-learn
boto3
joblib
python-dotenv